clean up code by breaking to smaller components
